% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-combine_iatf.R
\name{combine_iatf}
\alias{combine_iatf}
\title{Combine all Inter-Agency Task Force for the management of emerging
infectious diseases (IATF) resolutions datasets into a single dataset}
\source{
\url{https://www.doh.gov.ph/COVID-19/IATF-Resolutions}
}
\usage{
combine_iatf(res = 9:19)
}
\arguments{
\item{res}{A vector of resolution numbers to concatenate. Default is 9:19.}
}
\value{
A tibble of all the IATF resolutions
}
\description{
Combine all Inter-Agency Task Force for the management of emerging
infectious diseases (IATF) resolutions datasets into a single dataset
}
\examples{
combine_iatf()

}
