% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comotext.R
\name{comotext-deprecated}
\alias{comotext-deprecated}
\title{Deprecated functions in comotext}
\description{
These functions still work but will be removed in the next version.
}
\details{
\itemize{
  \item \code{\link{get_pr_url}}: This function is deprecated and will be
  removed in the next version of comotext.
  \item \code{\link{get_press_release}}: This function is deprecated and
  will be removed in the next version of comotext.
}
}
