% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfaLinks}
\alias{dfaLinks}
\title{Philippines Department of Foreign Affairs (DFA) press releases URLs}
\format{
A tibble with 2352 rows and 3 columns
\describe{
  \item{\code{id}}{Unique identifier}
  \item{\code{title}}{Title of press release}
  \item{\code{date}}{Date of issue of press release}
  \item{\code{source}}{Source of press release}
  \item{\code{type}}{Type of text data - press release}
  \item{\code{url}}{Absolute URLs of press releases}
}
}
\source{
\url{https://www.dfa.gov.ph/dfa-news/dfa-releasesupdate}. Updated
  as of 19 May 2020
}
\usage{
dfaLinks
}
\description{
Philippines Department of Foreign Affairs (DFA) press releases URLs
}
\examples{
dfaLinks

}
\keyword{datasets}
