% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_dfa.R
\name{get_dfa_links}
\alias{get_dfa_links}
\title{Get URLs of Philippines Department of Foreign Affairs (DFA) press releases}
\usage{
get_dfa_links(base = "https://www.dfa.gov.ph/dfa-news/dfa-releasesupdate")
}
\arguments{
\item{base}{Base URL for press releases in the Department of Foreign Affairs
website. Default is \url{https://www.dfa.gov.ph/dfa-news/dfa-releasesupdate}}
}
\value{
A tibble containing information on title of the press release,
  release date of press release and the absolute URL link to press release
}
\description{
Get URLs of Philippines Department of Foreign Affairs (DFA) press releases
}
\examples{
get_dfa_links(base = "http://www.dfa.gov.ph/dfa-news/dfa-releasesupdate")

}
