% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_doh.R
\name{get_doh_links}
\alias{get_doh_links}
\title{Extract title, URLs, unique identifiers and date of release of press releases
from the Department of Health website}
\usage{
get_doh_links(base = "https://www.doh.gov.ph/press-releases", pages = 1:25)
}
\arguments{
\item{base}{Base URL for press releases in the Department of Health website.
Default is \url{https://www.doh.gov.ph/press-releases}}

\item{pages}{A vector of page numbers corresponding to the page panel
containing the press release link. Default is 1:25. Press releases only go
up to page 25. Values higher than 25 will error.}
}
\value{
A tibble of 4 columns: 1) relative URLs of press release; 2) press
  release title; 3) press release unique identifier; and, 4) date of press
  release.
}
\description{
Extract title, URLs, unique identifiers and date of release of press releases
from the Department of Health website
}
\examples{
## Get press release URLs for first page of press release panel. The base
## argument has been specified using non-SSL version of base URL to prevent
## erroring for those with proxy server connections (i.e., in Travis tests)
get_doh_links(base = "http://www.doh.gov.ph/press-releases",
              pages = 1)

}
