% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_doh.R
\name{get_doh_release}
\alias{get_doh_release}
\title{Extract text of press release from the Philippines Department of Health
website}
\usage{
get_doh_release(df)
}
\arguments{
\item{df}{A data.frame created using \code{get_pr_url} providing values for
relative URL of press release/s, unique identifier of press release and,
date of issue of press release.}
}
\value{
A tibble containing text of the press release with additional
  information on line number, type of text, unique identifier and date of
  press release.
}
\description{
Extract text of press release from the Philippines Department of Health
website
}
\examples{
prURL <- get_doh_links(base = "http://www.doh.gov.ph/press-releases",
                       pages = 1)
get_doh_release(df = prURL[1, ])

}
