% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_iatf.R
\name{get_iatf_pdf}
\alias{get_iatf_pdf}
\title{Get a specific IATF resolution PDF file}
\usage{
get_iatf_pdf(id = iatfResList$id, ssl = TRUE)
}
\arguments{
\item{id}{A vector of resolution number/s of IATF resolution to be retrieved.
Numeric ranging from current resolution numbers available via
\code{iatfResList}. Default is all values for resolution identifiers
available via \code{iatfResList}}

\item{ssl}{Logical. Should SSL/https version of site be used? Default is
TRUE}
}
\value{
A temporary directory path pointing to a temporary file for PDF
  of IATF resolution required
}
\description{
Get a specific IATF resolution PDF file
}
\examples{
#get_iatf_pdf(id = 9, ssl = FALSE)


}
