% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_press_release.R
\name{get_pr_title}
\alias{get_pr_title}
\title{Extract title of press releases from Department of Health website}
\usage{
get_pr_title(pages = 1:13)
}
\arguments{
\item{pages}{A vector of page numbers corresponding to the page panel
containing the press release link}
}
\value{
A tibble of 2 columns: 1) press release title; and, 2) date of press
  release.
}
\description{
Extract title of press releases from Department of Health website
}
\examples{
get_pr_title(pages = 1:5)

}
