% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_press_release.R
\name{get_press_release}
\alias{get_press_release}
\title{Extract text of press release from the Philippines Department of Health
website}
\usage{
get_press_release(url, date)
}
\arguments{
\item{url}{of press release to extract text from}

\item{date}{Date press release was issued. Should be in <YYYY-MM-DD> format.}
}
\value{
A tibble containing text of the press release with additional
  information on line number, type of text and date of press release.
}
\description{
Extract text of press release from the Philippines Department of Health
website
}
\examples{
baseURL <- "http://www.doh.gov.ph"
prURL <- get_pr_url(pages = 1)[1, "url"]
get_press_release(url = paste(baseURL, prURL, sep = ""),
                  date = get_pr_url(pages = 1)[1, "date"])

}
