% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfResList}
\alias{iatfResList}
\title{IATF resolutions URL links}
\format{
A tibble with 25 rows and 4 columns:
\describe{
  \item{\code{id}}{Unique identifier}
  \item{\code{title}}{Title of resolution}
  \item{\code{date}}{Date resolution released. Format <YYYY-MM-DD>}
  \item{\code{link}}{Relative URL of IATF resolution}
}
}
\source{
\url{https://www.doh.gov.ph/COVID-19/IATF-Resolutions}. Updated
  as of 16 May 2020.
}
\usage{
iatfResList
}
\description{
IATF resolutions URL links
}
\examples{
iatfResList

}
\keyword{datasets}
