% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfResolution10}
\alias{iatfResolution10}
\title{COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 10}
\format{
A tibble with 84 rows and 2 columns:
\describe{
  \item{\code{linenumber}}{Linenumber}
  \item{\code{text}}{Character string of text about 70 characters}
}
}
\source{
\url{https://www.doh.gov.ph/sites/default/files/health-update/IATF-RESO-10.pdf}
}
\usage{
iatfResolution10
}
\description{
COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 10
}
\examples{
iatfResolution09

}
\keyword{datasets}
