% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfResolution12}
\alias{iatfResolution12}
\title{COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 12}
\format{
A tibble with 136 rows and 4 columns:
\describe{
  \item{\code{linenumber}}{Linenumber}
  \item{\code{text}}{Character string of text about 70 characters}
  \item{\code{resolution}}{Resolution number}
  \item{\code{date}}{Date in <YYYY-MM-DD> format when resolution was made}
}
}
\source{
\url{https://www.doh.gov.ph/sites/default/files/health-update/IATF-RESO-12.pdf}
}
\usage{
iatfResolution12
}
\description{
COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 12
}
\examples{
iatfResolution12

}
\keyword{datasets}
