% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfResolution15}
\alias{iatfResolution15}
\title{COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 15}
\format{
A tibble with 166 rows and 2 columns:
\describe{
  \item{\code{linenumber}}{Linenumber}
  \item{\code{text}}{Character string of text about 70 characters}
}
}
\source{
\url{https://www.doh.gov.ph/sites/default/files/health-update/IATF-RESO-15.pdf}
}
\usage{
iatfResolution15
}
\description{
COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 15
}
\examples{
iatfResolution15

}
\keyword{datasets}
