% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfResolution17}
\alias{iatfResolution17}
\title{COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 17}
\format{
A tibble with 151 rows and 5 columns:
\describe{
  \item{\code{linenumber}}{Linenumber}
  \item{\code{text}}{Character string of text about 70 characters}
  \item{\code{type}}{Type of document}
  \item{\code{id}}{Text document identifier}
  \item{\code{date}}{Date in <YYYY-MM-DD> format when resolution was made}
}
}
\source{
\url{https://www.doh.gov.ph/sites/default/files/health-update/IATF Resolution No. 17.pdf}
}
\usage{
iatfResolution17
}
\description{
COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 17
}
\examples{
iatfResolution17

}
\keyword{datasets}
