% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prLinks}
\alias{prLinks}
\title{Department of Health press releases URLs from 2017 to May 2020}
\format{
A tibble with 353 rows and 3 columns:
\describe{
  \item{\code{url}}{Relative URL of press releases}
  \item{\code{id}}{Unique identifier of press release}
  \item{\code{date}}{Date of issue of press release}
}
}
\source{
Produced by \code{get_pr_url(pages = 1:25)}. Updated as of
  10 May 2020
}
\usage{
prLinks
}
\description{
Department of Health press releases URLs from 2017 to May 2020
}
\examples{
prLinks

}
\keyword{datasets}
