% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_press_release.R
\name{get_press_release}
\alias{get_press_release}
\title{Extract text of press release from the Philippines Department of Health
website}
\usage{
get_press_release(url)
}
\arguments{
\item{url}{of press release to extract text from}
}
\value{
A tibble containing text of the press release with additional
  information on line number, type of text and date of press release.
}
\description{
Extract text of press release from the Philippines Department of Health
website
}
\examples{
baseURL <- "https://www.doh.gov.ph/press-release"
prTitle <- "ECQ-Buys-PH-Time-Continued-Practice-of-Healthy-Behavior-Needed-to-Sustain-Gains"
get_press_release(url = paste(baseURL, prTitle, sep = "/"))

}
