% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_press_release.R
\name{get_press_release}
\alias{get_press_release}
\title{Extract text of press release from the Philippines Department of Health
website}
\usage{
get_press_release(base = "https://www.doh.gov.ph", df)
}
\arguments{
\item{base}{Base URL for press releases in the Department of Health website.
Default is \url{https://www.doh.gov.ph}}

\item{df}{A data.frame created using \code{get_pr_url} providing values for
relative URL of press release/s, unique identifier of press release and,
date of issue of press release.}
}
\value{
A tibble containing text of the press release with additional
  information on line number, type of text, unique identifier and date of
  press release.
}
\description{
Extract text of press release from the Philippines Department of Health
website
}
\examples{
prURL <- get_pr_url(base = "http://www.doh.gov.ph/press-releases",
                    pages = 1)
get_press_release(base = "http://www.doh.gov.ph",
                  df = prURL[1, ])

}
