% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_press_release.R
\name{get_press_release}
\alias{get_press_release}
\title{Extract text of press release from the Philippines Department of Health
website}
\usage{
get_press_release(base = "https://www.doh.gov.ph", url, date)
}
\arguments{
\item{base}{Base URL for press releases in the Department of Health website.
Default is \url{https://www.doh.gov.ph}}

\item{url}{of press release to extract text from}

\item{date}{Date press release was issued. Should be in <YYYY-MM-DD> format.}
}
\value{
A tibble containing text of the press release with additional
  information on line number, type of text and date of press release.
}
\description{
Extract text of press release from the Philippines Department of Health
website
}
\examples{
prURL <- get_pr_url(pages = 1)
get_press_release(base = "http://www.doh.gov.ph",
                  url = prURL$url[1],
                  date = prURL$date[1])

}
